e2e
Update the store and tests
