+++ matrix-sdk/src/main/java/org/matrix/androidsdk/data/MXFileCryptoStoreMetaData.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2015 OpenMarket Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.matrix.androidsdk.data;
+
+public class MXFileCryptoStoreMetaData implements java.io.Serializable {
+    // The obtained user id.
+    public String mUserId;
+
+    // the device id
+    public String mDeviceId;
+
+    //  The current version of the store.
+    public int mVersion;
+
+    // flag to tell if the device is announced
+    public boolean mDeviceAnnounced;
+}
